Refactored ConfigureComponent and the databasesynchronizer classes
