+++ viewer-admin/src/test/resources/log4j.xml
@@ -16,7 +16,7 @@ LEVELS: debug, info, warn, error, fatal en off, all -->
     <logger name="nl.b3p.viewer.config.services">
         <level value="info" />
     </logger>
-    <logger name="nl.b3p.viewer.util.databaseupdate.ScriptRunner">
+    <logger name="nl.b3p.viewer.util.ScriptRunner">
         <level value="error" />
     </logger>
     <logger name="org.hibernate">
