+++ viewer-config-persistence-helpers/src/main/java/nl/b3p/viewer/config/stripersist/DynamicStripersistInitializer.java
