+++ viewer-config-persistence-helpers/src/test/java/nl/b3p/viewer/util/ViewerConfigPersistenceSuite.java
@@ -16,8 +16,8 @@
  */
 package nl.b3p.viewer.util;
 
-import databaseupdate.DatabaseSynchronizerEMTest;
-import databaseupdate.DatabaseSynchronizerTest;
+import nl.b3p.viewer.util.databaseupdate.DatabaseSynchronizerEMTest;
+import nl.b3p.viewer.util.databaseupdate.DatabaseSynchronizerTest;
 import nl.b3p.viewer.config.app.BookmarkTest;
 import nl.b3p.viewer.config.app.StartLayerTest;
 import nl.b3p.viewer.config.app.StartLevelTest;
