+++ viewer-config-persistence/pom.xml
@@ -195,15 +195,6 @@
             <artifactId>json</artifactId>
         </dependency>
         <dependency>
-            <groupId>org.flamingo-mc</groupId>
-            <artifactId>viewer-commons</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.flamingo-mc</groupId>
-            <artifactId>web-commons</artifactId>
-        </dependency>
-
-        <dependency>
             <groupId>org.javassist</groupId>
             <artifactId>javassist</artifactId>
         </dependency>
