+++ viewer/src/test/resources/log4j.xml
@@ -34,7 +34,7 @@
     <logger name="httpclient.wire">
         <level value="info" />
     </logger>
-    <logger name="nl.b3p.viewer.util.databaseupdate.ScriptRunner">
+    <logger name="nl.b3p.viewer.util.ScriptRunner">
         <level value="error" />
     </logger>
     <logger name="nl.b3p.viewer.userlayer">
